[seller] refactor(171): 리뷰반영(FSD원칙에 따라 상수 분리)
