+++ apps/seller/src/entity/register/index.ts
@@ -8,11 +8,7 @@ export {
   verifyAccount,
 } from './register.api'
 export { registerQueries } from './register.query'
-export {
-  FILE_UPLOAD_LIMITS,
-  REGISTER_MESSAGES,
-  REGISTER_TOAST_MESSAGES,
-} from './register.constant'
+export { FILE_UPLOAD_LIMITS } from './register.constant'
 export {
   registerSchema,
   storeInfoSchema,
