+++ apps/seller/src/pages/register/index.ts
@@ -1,0 +1,1 @@
+export { REGISTER_MESSAGES } from './register.constant'
