+++ apps/seller/src/pages/register/register.constant.ts
@@ -1,0 +1,7 @@
+export const REGISTER_MESSAGES = {
+  COMPLETE: {
+    TITLE: '🎉 회원가입 완료',
+    DESCRIPTION:
+      '빵그리의 오븐 판매자 채널에 가입해 주셔서 진심으로 감사드립니다. 고객님의 가입 정보는 확인 후 승인 절차를 거치게 되며, 승인까지는 영업일 기준 1~2일 정도 소요됩니다. 승인될 때까지 조금만 기다려 주세요.',
+  },
+} as const
